move dates -> datetime use utcnow with timezone
